+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
 
 ## 0.0.1-dev
+- Added decimalPlaces parameter to round and floor.
 - Initial